refactor(stdlib): extract shared template_utils from map and template
Extract getByPath and renderTemplate into template_utils.ts. Fix ReDoS
vulnerability in template regex and add Object.hasOwn guard to block
prototype chain traversal in getByPath.
