+++ src/commands/stdlib/map.ts
@@ -1,23 +1,4 @@
-function getByPath(obj: any, path: string): any {
-  if (path === '.' || path === 'this') return obj;
-  const parts = path.split('.').filter(Boolean);
-  let cur: any = obj;
-  for (const p of parts) {
-    if (cur == null) return undefined;
-    cur = cur[p];
-  }
-  return cur;
-}
-
-function renderTemplate(tpl: string, ctx: any): string {
-  return tpl.replace(/\{\{\s*([^}]+?)\s*\}\}/g, (_m, expr) => {
-    const key = String(expr ?? '').trim();
-    const val = getByPath(ctx, key);
-    if (val === undefined || val === null) return '';
-    if (typeof val === 'string') return val;
-    return JSON.stringify(val);
-  });
-}
+import { renderTemplate } from './template_utils.js';
 
 function parseAssignments(tokens: any[]): Array<{ key: string; value: string }> {
   const out: Array<{ key: string; value: string }> = [];
